extract function: normalizeFeed
